ILS.pm: ILS decides how to handle items on hold
Change behaviour for items on loan: orignal code denies checkouts for
items on loan by explicitly checking the hold queue. Arguably, this
descision should be up to the ILS and thus be configurable. join2 e.g.
checks out with a short loan period in case of requests also to handle
book transports transparently.

Closes join2/join2-data#2170